REF: fillna tests (#33183)
